+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog (nionswift-tool)
 ==========================
 
+0.4.5 (2020-09-03)
+------------------
+- Add function to allow handling of window closing and quit.
+- Add function to truncate string to pixel width.
+
 0.4.4 (2020-05-08)
 ------------------
 - Add color/font methods to eliminate need for stylesheet properties.